refactor: css 패키지 리팩토링 (#13)
