+++ src/App.tsx
@@ -1,5 +1,5 @@
 import MainRouter from './router/MainRouter'
-import './App.css'
+import './styles/App.css'
 
 function App() {
   return <MainRouter />
