+++ src/index.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom/client'
-import './index.css'
+import './styles/index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import {RecoilRoot} from 'recoil'
