+++ src/styles/board.css
